demo-eng/adk_agents/adk_common/pyproject.toml
[project]
name = "adk_common"
version = "3.20251118.1"
description = "Common utilities for Tahoe agents"
authors = [{ name = "Gonzalo Varela", email = "gvarelal@google.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk==1.24.1",
    "python-dotenv==1.1.1",
    "google-genai==1.62.0",
    "google-cloud-aiplatform[adk,agent-engines]==1.136.0",
    "pydantic==2.11.7",
    "absl-py==2.3.1",
    "requests",
    "google-cloud-storage",
    "tenacity",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["adk_common"]
